requirements.txt
fastapi
uvicorn
line-bot-sdk
requests
python-dotenv
SQLAlchemy
psycopg2-binary